+++ src/persistence/PackageExporter.js
@@ -1,4 +1,5 @@
 import { SettingsManager } from './SettingsManager.js';
+import { collectSoundReferences } from './soundReferences.js';
 import { LocalBackend } from '../api/LocalBackend.js';
 import { collectRuntimeModules } from './runtimeModules.js';
 import { toFileSlug } from '../core/utils/filenames.js';
@@ -41,7 +42,7 @@ export const PackageExporter = {
 			}
 			const zip = new JSZip();
 
-			const soundFiles = this.collectSoundFiles(buzzData);
+			const soundFiles = collectSoundReferences(buzzData);
 
 			await this.addSoundFiles(zip, soundFiles);
 
@@ -110,62 +111,6 @@ export const PackageExporter = {
 			console.error('Package export failed:', error);
 			throw error;
 		}
-	},
-
-	collectSoundFiles(buzzData) {
-		const soundFiles = new Set();
-
-		if (buzzData.sounds) {
-			buzzData.sounds.forEach(sound => {
-				if (sound.type === 'SoundFile' && sound.params?.soundFile) {
-					soundFiles.add(sound.params.soundFile);
-				}
-
-				if (sound.type === 'Sampler') {
-					if (sound.params?.samplerMode === 'single' && sound.params?.soundFile) {
-						soundFiles.add(sound.params.soundFile);
-					}
-
-					if (sound.params?.samplerMode === 'grid' && sound.params?.gridSamples) {
-						Object.values(sound.params.gridSamples).forEach(sample => {
-							if (sample.fileName) {
-								soundFiles.add(sample.fileName);
-							}
-						});
-					}
-				}
-			});
-		}
-
-		if (buzzData.sequencers) {
-			buzzData.sequencers.forEach(sequencer => {
-				if (sequencer.tracks) {
-					sequencer.tracks.forEach(track => {
-						if (track.instrumentType === 'synth' && track.synthParams) {
-							if (track.synthType === 'SoundFile' && track.synthParams.soundFile) {
-								soundFiles.add(track.synthParams.soundFile);
-							}
-
-							if (track.synthType === 'Sampler') {
-								if (track.synthParams.samplerMode === 'single' && track.synthParams.soundFile) {
-									soundFiles.add(track.synthParams.soundFile);
-								}
-
-								if (track.synthParams.samplerMode === 'grid' && track.synthParams.gridSamples) {
-									Object.values(track.synthParams.gridSamples).forEach(sample => {
-										if (sample.fileName) {
-											soundFiles.add(sample.fileName);
-										}
-									});
-								}
-							}
-						}
-					});
-				}
-			});
-		}
-
-		return Array.from(soundFiles);
 	},
 
 	async addSoundFiles(zip, soundFiles) {